\name{X13CMS-package}
\alias{X13CMS-package}
\alias{X13CMS}
\docType{package}
\title{
Global tracking of isotopic labels in untargeted metabolomics
}
\description{
X13CMS extends XCMS's functionality to detect isotopologue groups in untargeted LC/MS profiling data that are enriched for isotope labels derived from labeled precursors. It also performs differential labeling analysis to reveal changes in labeling patterns under different biologial conditions.
}
\details{
\tabular{ll}{
Package: \tab X13CMS\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2014-12-07\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
Xiaojing Huang

Maintainer: Xiaojing Huang <huangxi@wusm.wustl.edu>
}
\references{
Huang, X et al. (2014) "X13CMS: Global Tracking of Isotopic Labels in Untargeted Metabolomics". Anal. Chem. 86, 1632−1639.
}

\keyword{ package }
\seealso{

}
\examples{

}
